Parse multiple errors (#1)
Previous to this the parser stage of the compiler would immediately exit on any error encountered. This has now been changed so that the parser continues to parse for as long as it can even if it encounters bad/unexpected tokens (early end-of-file is still, obviously, an immediate exit). All of this to say: The compiler camn now output multiple errors (and warnings) when parsing a file.